Move search-past-grants-alpha URL
