+++ controllers/aliases.js
@@ -47,8 +47,8 @@ const aliases = {
     '/freedom-of-information': '/about/customer-service/freedom-of-information',
     '/funded-projects': '/funding/past-grants',
     '/funding-uk': '/funding',
-    '/funding/awards-for-all': '/funding/under10k',
     '/funding/Awards-For-All': '/funding/under10k',
+    '/funding/awards-for-all': '/funding/under10k',
     '/funding/funding-guidance': '/funding',
     '/funding/funding-guidance/applying-for-funding': '/funding',
     '/funding/funding-guidance/applying-for-funding/help-using-our-electronic-application-forms': '/funding-guidance/help-using-our-application-forms',
@@ -61,6 +61,7 @@ const aliases = {
     '/funding/funding-guidance/managing-your-funding/self-evaluation': '/funding/funding-guidance/managing-your-funding/evaluation',
     '/funding/programmes/digital-funding': '/funding/programmes/digital-fund',
     '/funding/scotland-portfolio': '/funding/programmes?location=scotland',
+    '/funding/search-past-grants-alpha': '/funding/past-grants/search',
     '/global-content/programmes/england/awards-for-all-england': '/funding/programmes/national-lottery-awards-for-all-england',
     '/global-content/programmes/england/building-better-opportunities': '/funding/programmes/building-better-opportunities',
     '/global-content/programmes/england/building-better-opportunities/building-better-opportunities-resources': '/funding/programmes/building-better-opportunities/building-better-opportunities-resources',
