--- controllers/funding/views/past-grants.njk
@@ -1,71 +1,0 @@
-{% from "components/hero.njk" import hero with context %}
-{% from "components/breadcrumb-trail/macro.njk" import breadcrumbTrail %}
-{% from "components/feedback.njk" import inlineFeedback with context %}
-{% from "components/promo-card/macro.njk" import promoCard %}
-
-{% extends "layouts/main.njk" %}
-
-{% block content %}
-    <main role="main">
-        {{ hero(
-            titleText = copy.title,
-            image = heroImage,
-            accent = pageAccent
-        ) }}
-
-        <div class="nudge-up">
-            <div class="content-box u-inner-wide-only accent--{{ pageAccent }} a--border-top">
-                {{ breadcrumbTrail(breadcrumbs) }}
-
-                <section class="s-prose u-constrained-content-wide">
-                    <p>{{ copy.intro }}</p>
-                    {{ inlineFeedback(
-                        description = "Past Grants",
-                        promptLabel = copy.survey.prompt,
-                        fieldLabel = copy.survey.label
-                    )  }}
-                    <hr />
-                    <p>{{ copy.optionsIntro }}</p>
-                </section>
-
-                <ul class="flex-grid">
-                    <li class="flex-grid__item">
-                        <div class="card">
-                            <header class="card__header">
-                                <h2 class="card__title">{{ copy.grantnav.title }}</h2>
-                            </header>
-                            <div class="card__body">
-                                <p><a
-                                    data-ga-on="click"
-                                    data-ga-event-category="Past grants"
-                                    data-ga-event-action="GrantNav"
-                                    class="btn btn--medium"
-                                    href="{{ grantNavLink }}">
-                                    {{ copy.grantnav.linkText }}
-                                </a><p>
-                                <p>{{ copy.grantnav.subtitle }}</p>
-                            </div>
-                        </div>
-                    </li>
-                    <li class="flex-grid__item">
-                        <div class="card">
-                            <header class="card__header">
-                                <h2 class="card__title">{{ copy.legacy.title }}</h2>
-                            </header>
-                            <div class="card__body">
-                                <p><a
-                                    data-ga-on="click"
-                                    data-ga-event-category="Past grants"
-                                    data-ga-event-action="Legacy past grants"
-                                    class="btn btn--medium"
-                                    href="{{ localify('/funding/search-past-grants') }}">
-                                    {{ copy.legacy.linkText }}
-                                </a></p>
-                            </div>
-                        </div>
-                    </li>
-                </ul>
-            </div>
-        </div>
-    </main>
-{% endblock %}
