+++ controllers/past-grants/views/recipient-detail.njk
@@ -90,7 +90,7 @@
                         <div class="card u-no-margin">
                             <header class="card__header">
                                 <h3 class="card__title u-no-margin">
-                                    <a href="{{ localify("/funding/search-past-grants-alpha/grant/" + grant.id) }}">
+                                    <a href="{{ localify("/funding/past-grants/search/grant/" + grant.id) }}">
                                         {{ grant.title }}
                                     </a>
                                 </h3>
